--- .vscode/launch.json
@@ -1,18 +1,0 @@
-{
-  // Use IntelliSense to learn about possible attributes.
-  // Hover to view descriptions of existing attributes.
-  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "name": "Debug Terraform Provider",
-      "type": "go",
-      "request": "launch",
-      "mode": "debug",
-      // this assumes your workspace is the root of the repo
-      "program": "${workspaceFolder}",
-      "env": {},
-      "args": ["-debug"]
-    }
-  ]
-}
