--- .vscode/tasks.json
@@ -1,19 +1,0 @@
-{
-	"version": "2.0.0",
-	"tasks": [
-		{
-			"type": "shell",
-			"label": "Exec sg-tf-provider",
-			// Env Var STACKGUARDIAN_API_KEY is required
-			"command": "if not set -q STACKGUARDIAN_API_KEY; echo 'Env Var STACKGUARDIAN_API_KEY is required'; else; rm -rfv .terraform .terraform.lock.hcl && pushd ../..; if not make; popd && echo 'BUILD KO'; else; popd && echo 'BUILD OK' && terraform init && TF_LOG= terraform plan; TF_LOG= terraform apply; end; end;",
-			"options": {
-				"cwd": "examples/workflow_outputs_example"
-			},
-			"group": {
-				"kind": "test",
-				"isDefault": true
-			},
-			"problemMatcher": []
-		}
-	]
-}
