+++ internal/datasources/workflow_group/schema.go
@@ -1,17 +1,16 @@
-package workflowGroups
+package workflowgroupdatasource
 
 import (
 	"context"
 	"fmt"
 
 	"github.com/StackGuardian/terraform-provider-stackguardian/internal/constants"
-	"github.com/hashicorp/terraform-plugin-framework/resource"
-	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/datasource"
+	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
 
-// Schema defines the schema for the resource.
-func (r *workflowGroupResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
+func (d *workflowGroupDataSource) Schema(_ context.Context, _ datasource.SchemaRequest, resp *datasource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		Attributes: map[string]schema.Attribute{
 			"resource_name": schema.StringAttribute{
@@ -20,16 +19,13 @@ func (r *workflowGroupResource) Schema(_ context.Context, _ resource.SchemaReque
 			},
 			"description": schema.StringAttribute{
 				MarkdownDescription: fmt.Sprintf(constants.Description, "workflow group"),
-				Optional:            true,
 				Computed:            true,
 			},
 			"tags": schema.ListAttribute{
 				MarkdownDescription: fmt.Sprintf(constants.Tags, "workflow group"),
 				ElementType:         types.StringType,
-				Optional:            true,
 				Computed:            true,
 			},
 		},
 	}
-
 }
