+++ internal/provider/provider.go
@@ -1,4 +1,4 @@
-package provider
+package sgprovider
 
 import (
 	"context"
@@ -10,15 +10,15 @@ import (
 	"github.com/StackGuardian/terraform-provider-stackguardian/internal/customTypes"
 	connectordatasource "github.com/StackGuardian/terraform-provider-stackguardian/internal/datasources/connector"
 	roledatasource "github.com/StackGuardian/terraform-provider-stackguardian/internal/datasources/role"
-	roleassignment "github.com/StackGuardian/terraform-provider-stackguardian/internal/datasources/roleAssignment"
+	roleassignmentdatasource "github.com/StackGuardian/terraform-provider-stackguardian/internal/datasources/roleassignment"
 	stackoutputs "github.com/StackGuardian/terraform-provider-stackguardian/internal/datasources/stack_outputs"
 	stackworkflowoutputs "github.com/StackGuardian/terraform-provider-stackguardian/internal/datasources/stack_workflow_outputs"
-	workflowgroupsdatasource "github.com/StackGuardian/terraform-provider-stackguardian/internal/datasources/workflow_groups"
+	workflowgroupdatasource "github.com/StackGuardian/terraform-provider-stackguardian/internal/datasources/workflow_group"
 	workflowoutputs "github.com/StackGuardian/terraform-provider-stackguardian/internal/datasources/workflow_outputs"
 	"github.com/StackGuardian/terraform-provider-stackguardian/internal/resource/connector"
 	"github.com/StackGuardian/terraform-provider-stackguardian/internal/resource/role"
-	"github.com/StackGuardian/terraform-provider-stackguardian/internal/resource/roleAssignment"
-	"github.com/StackGuardian/terraform-provider-stackguardian/internal/resource/workflowGroups"
+	roleassignment "github.com/StackGuardian/terraform-provider-stackguardian/internal/resource/role_assignment"
+	workflowgroup "github.com/StackGuardian/terraform-provider-stackguardian/internal/resource/workflow_group"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/path"
 	"github.com/hashicorp/terraform-plugin-framework/provider"
@@ -204,8 +204,8 @@ func (p *stackguardianProvider) DataSources(_ context.Context) []func() datasour
 		stackworkflowoutputs.NewDataSource,
 		workflowoutputs.NewDataSource,
 		connectordatasource.NewDataSource,
-		roleassignment.NewDataSource,
-		workflowgroupsdatasource.NewDataSource,
+		roleassignmentdatasource.NewDataSource,
+		workflowgroupdatasource.NewDataSource,
 		roledatasource.NewDataSource,
 	}
 }
@@ -214,8 +214,8 @@ func (p *stackguardianProvider) DataSources(_ context.Context) []func() datasour
 func (p *stackguardianProvider) Resources(_ context.Context) []func() resource.Resource {
 	return []func() resource.Resource{
 		connector.NewResource,
-		workflowGroups.NewResource,
+		workflowgroup.NewResource,
 		role.NewResource,
-		roleAssignment.NewResource,
+		roleassignment.NewResource,
 	}
 }
