+++ internal/resource/role_assignment/model.go
@@ -1,4 +1,4 @@
-package roleAssignment
+package roleassignment
 
 import (
 	"context"
