+++ internal/resource/workflow_group/schema.go
@@ -1,16 +1,17 @@
-package workflowgroupsdatasource
+package workflowgroup
 
 import (
 	"context"
 	"fmt"
 
 	"github.com/StackGuardian/terraform-provider-stackguardian/internal/constants"
-	"github.com/hashicorp/terraform-plugin-framework/datasource"
-	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
+	"github.com/hashicorp/terraform-plugin-framework/resource"
+	"github.com/hashicorp/terraform-plugin-framework/resource/schema"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 )
 
-func (d *workflowGroupsDataSource) Schema(_ context.Context, _ datasource.SchemaRequest, resp *datasource.SchemaResponse) {
+// Schema defines the schema for the resource.
+func (r *workflowGroupResource) Schema(_ context.Context, _ resource.SchemaRequest, resp *resource.SchemaResponse) {
 	resp.Schema = schema.Schema{
 		Attributes: map[string]schema.Attribute{
 			"resource_name": schema.StringAttribute{
@@ -19,13 +20,16 @@ func (d *workflowGroupsDataSource) Schema(_ context.Context, _ datasource.Schema
 			},
 			"description": schema.StringAttribute{
 				MarkdownDescription: fmt.Sprintf(constants.Description, "workflow group"),
+				Optional:            true,
 				Computed:            true,
 			},
 			"tags": schema.ListAttribute{
 				MarkdownDescription: fmt.Sprintf(constants.Tags, "workflow group"),
 				ElementType:         types.StringType,
+				Optional:            true,
 				Computed:            true,
 			},
 		},
 	}
+
 }
