+++ main.go
@@ -5,7 +5,7 @@ import (
 	"flag"
 	"log"
 
-	"github.com/StackGuardian/terraform-provider-stackguardian/internal/provider"
+	sgprovider "github.com/StackGuardian/terraform-provider-stackguardian/internal/provider"
 	"github.com/hashicorp/terraform-plugin-framework/providerserver"
 )
 
@@ -27,7 +27,7 @@ func main() {
 		Debug:   debug,
 	}
 
-	err := providerserver.Serve(context.Background(), provider.New(Version), opts)
+	err := providerserver.Serve(context.Background(), sgprovider.New(Version), opts)
 
 	if err != nil {
 		log.Fatal(err.Error())
